--- SigningServer.Client/NLog.config
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      autoReload="true"
-      throwExceptions="false"
-      internalLogLevel="Off" internalLogFile="c:\temp\nlog-internal.log" >
-
-  <targets>
-    <target xsi:type="ColoredConsole" name="console" layout="${longdate} ${level} - ${message} ${exception:format=ToString}"/>
-    <target xsi:type="Debugger" name="debugger" layout="${longdate} ${level} - ${message} ${exception:format=ToString}"/>
-  </targets>
-
-  <rules>
-    <logger name="*" minlevel="Trace" writeTo="console,debugger" />
-  </rules>
-</nlog>
