+++ SigningServer.Client/appsettings.json
@@ -1,0 +1,17 @@
+﻿{
+    "NLog": {
+      "targets": {
+        "console": {
+          "type": "ColoredConsole",
+          "layout": "${longdate} ${level} - ${message} ${exception:format=ToString}"
+        }
+      },
+      "rules": [
+        {
+          "logger": "*",
+          "minLevel": "Trace",
+          "writeTo": "console"
+        }
+      ]
+    }
+}
