+++ SigningServer.Server/appsettings.json
@@ -21,5 +21,35 @@
         }
       }
     ]
+  },
+  "NLog": {
+    "targets": {
+      "async": true,
+      "console": {
+        "type": "ColoredConsole",
+        "layout": "${longdate} ${level} ${callsite:cleanNamesOfAsyncContinuations=true:cleanNamesOfAnonymousDelegates=true} - ${message} ${exception:format=ToString}"
+      },
+      "file": {
+        "type": "File",
+        "layout": "${longdate} ${level} ${callsite:cleanNamesOfAsyncContinuations=true:cleanNamesOfAnonymousDelegates=true} - ${message} ${exception:format=ToString}",
+        "archiveEvery": "Day",
+        "archiveNumbering": "DateAndSequence",
+        "archiveFileName": "${basedir}/signingserver.{#}.log",
+        "archiveDateFormat": "yyyy-MM-dd",
+        "archiveAboveSize": 52480000,
+        "maxArchiveFiles": 10,
+        "keepFileOpen": true,
+        "concurrentWrites": true,
+        "openFileCacheTimeout": 30
+      }
+    },
+    "rules": [
+      {
+        "enabled": true,
+        "logger": "*",
+        "minLevel": "Trace",
+        "writeTo": "console,file"
+      }
+    ]
   }
 }
