--- SigningServer.Server/nlog.config
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      autoReload="true"
-      throwExceptions="false"
-      internalLogLevel="Off" internalLogFile="c:\temp\nlog-internal.log" >
-  <targets>
-    <target xsi:type="ColoredConsole" name="console" layout="${longdate} ${level} ${callsite:cleanNamesOfAsyncContinuations=true:cleanNamesOfAnonymousDelegates=true} - ${message} ${exception:format=ToString}"/>
-      <target name="file" xsi:type="AsyncWrapper" queueLimit="5000" overflowAction="Discard">
-          <target xsi:type="File" name="f" fileName="${basedir}/signingserver.log"
-                  layout="${longdate} ${level} ${callsite:cleanNamesOfAsyncContinuations=true:cleanNamesOfAnonymousDelegates=true} - ${message} ${exception:format=ToString}"
-                  archiveEvery="Day"
-                  archiveNumbering="DateAndSequence"
-                  archiveFileName="${basedir}/signingserver.{#}.log"
-                  archiveDateFormat="yyyy-MM-dd"
-                  archiveAboveSize="52480000"
-                  maxArchiveFiles="10"
-                  keepFileOpen="true"
-                  concurrentWrites="true"
-                  openFileCacheTimeout="30"
-          />
-      </target>
-    <target xsi:type="Debugger" name="debugger" layout="${longdate} ${level} ${callsite:cleanNamesOfAsyncContinuations=true:cleanNamesOfAnonymousDelegates=true} - ${message} ${exception:format=ToString}"/>
-  </targets>
-
-  <rules>
-    <logger name="*" minlevel="Trace" writeTo="console,file,debugger" />
-  </rules>
-</nlog>
