+++ SigningServer.Signing/SigningServer.Signing.csproj
@@ -1,0 +1,8 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <ItemGroup>
+    <ProjectReference Include="..\SigningServer.Android\SigningServer.Android.csproj" />
+    <ProjectReference Include="..\SigningServer.ClickOnce\SigningServer.ClickOnce.csproj" />
+    <ProjectReference Include="..\SigningServer.MsSign\SigningServer.MsSign.csproj" />
+    <ProjectReference Include="..\SigningServer.NuGet\SigningServer.NuGet.csproj" />
+  </ItemGroup>
+</Project>
